video/requirements.txt
ffmpeg-python>=0.2.0
Pillow>=10.0.0
pytest>=7.4.0
